I8C: Reject "deref int" and derivatives

+++ tests/test_deref_offset_base.py
@@ -27,7 +27,7 @@ SOURCE = """\
 define test::deref_offset_base returns int
    extern ptr base
    argument int offset
-   deref offset(base), int
+   deref offset(base), intptr_t
 """
 
 class TestDerefOffsetBase(TestCase):
